split vm2 functions to seperate file
* remove some unused functions
